[Refactor] Move the Config Struct from cmd/main.go to pkg/config
